+++ packages/plugins/htpasswd/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 14.0.0-next-9.6
+
+### Patch Changes
+
+- Updated dependencies [1905990]
+  - @verdaccio/core@9.0.0-next-9.6
+  - @verdaccio/file-locking@14.0.0-next-9.2
+
 ## 14.0.0-next-9.5
 
 ### Patch Changes