+++ packages/plugins/ui-theme/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @verdaccio/ui-theme
 
+## 9.0.0-next-9.6
+
+### Patch Changes
+
+- 80ddcda: chore(ui): remove babel dependency
+- 9d5360b: chore: vite replace rollupOptions with rolldownOptions
+
 ## 9.0.0-next-9.5
 
 ### Patch Changes