+++ packages/search/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @verdaccio/search
 
+## 9.0.0-next-9.6
+
+### Patch Changes
+
+- Updated dependencies [1905990]
+  - @verdaccio/core@9.0.0-next-9.6
+  - @verdaccio/config@9.0.0-next-9.6
+  - @verdaccio/logger@9.0.0-next-9.6
+  - @verdaccio/proxy@9.0.0-next-9.6
+
 ## 9.0.0-next-9.5
 
 ### Patch Changes